backend/requirements.txt
flask==3.1.2
openai==2.14.0
python-dotenv==1.0.1
